Adds MSVC 2010 (x86 and x64) dependencies for CorsixTH 0.40
This commit includes all dependencies needed to build CorsixTH 0.40 in
MSVC 2010. They will likely work in later versions of Visual Studio as well.

For the most part the versions are those included with the official release,
with the exception of Lua where Lua 5.1.4 is included instead of LuaJIT 2.0.3.
msinttypes is not included because currently CMake cannot find them in the
CMAKE_INCLUDE_PATH and they are header files which do not need to be built.

Includes:
ffmpeg-2.2.3 (lgpl build)
freetype-2.5.3
lua 5.1.4
luasocket-2013-10-24 (git)
sdl-1.2.15
sdl_mixer-1.2.12
wxWidgets-2.9.4